Filter fields by blacklist instead of whitelist
This allows sharing of the cache between multiple runs of monitor with
different sets of queries. We only blacklist the "description" field
(there are probably more fields we don't care about from a transition
point of view, but "description" is the most verbose one).

As a consequence, do not rely on the presence of "edos-debcheck" in
the query anymore; add an explicit command-line option instead. A
posteriori, it looks cleaner this way anyway.