Handle the media dir automatically
When ben creates a tracker or a monitor, it should put a symlink to
the "media/" directory that contains all necessary materials for the
generated web page.

For the monitor, ben puts a symlink only when given a filename for the
final location of the monitor. Otherwise, we have no correct way to
know when the symlink should be created.

If a "media" element on disk is present, we skip this step.

By default, the media dir is "/usr/share/ben/media" but the user can
still overwrite this default by setting the environement variable
"BEN_MEDIA_DIR". If a "media" symlink is present and doesn't point to
the registered media dir, the symlink is recreated.